pyproject.toml
[project]
name = "mooc"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "django (==5.2)",
    "psycopg (>=3.2.9,<4.0.0)",
    "asgiref (>=3.8.1,<4.0.0)",
    "authlib (>=1.6.0,<2.0.0)",
    "certifi (>=2025.6.15,<2026.0.0)",
    "charset-normalizer (>=3.4.2,<4.0.0)",
    "coverage (>=7.9.1,<8.0.0)",
    "django-browser-reload (>=1.18.0,<2.0.0)",
    "django-debug-toolbar (>=5.2.0,<6.0.0)",
    "django-tailwind (>=4.0.1,<5.0.0)",
    "exceptiongroup (>=1.3.0,<2.0.0)",
    "faker (>=37.4.0,<38.0.0)",
    "idna (>=3.10,<4.0)",
    "iniconfig (>=2.1.0,<3.0.0)",
    "packaging (>=25.0,<26.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "pluggy (>=1.6.0,<2.0.0)",
    "pycparser (>=2.22,<3.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "pytest-cov (>=6.2.1,<7.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "sqlparse (>=0.5.3,<0.6.0)",
    "tomli (>=2.2.1,<3.0.0)",
    "typing-extensions (>=4.14.0,<5.0.0)",
    "urllib3 (>=2.5.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
